requirements.txt
Django~=3.2.4
sentry-sdk~=1.1.0
django-simple-captcha~=0.5.14
django-photologue~=3.13
django-crispy-forms~=1.12.0
fontawesomefree==6.0.0